chore(builder): deprecate snippets implementation
