+++ libs/builder/helpers/snippets-from-asset.ts
@@ -2,6 +2,7 @@ import {NgDocStyleType} from '@ng-doc/core';
 
 import {NgDocAsset} from '../interfaces';
 import {codeTypeFromExt} from './code-type-from-ext';
+import {processLegacySnippets} from './process-legacy-snippets';
 import {processSnippets} from './process-snippets';
 
 /**
@@ -11,7 +12,7 @@ import {processSnippets} from './process-snippets';
  * @param inlineStylesType - Inline styles type
  */
 export function snippetsFromAsset(asset: NgDocAsset, inlineStylesType: NgDocStyleType): NgDocAsset[] {
-	const snippets = processSnippets(asset.code);
+	const snippets = processLegacySnippets(asset.code).concat(processSnippets(asset.code));
 	const codeType = codeTypeFromExt(asset.filePath);
 	const isStylesFile = ['CSS', 'SCSS', 'LESS', 'SASS'].includes(codeType.toUpperCase());
